extract pm quality review routes
